+++ package.json
@@ -2,7 +2,7 @@
   "name": "@npm/types",
   "version": "1.0.0",
   "description": "Typescript definitions for npm registry content",
-  "types": "src/index.d.ts",
+  "types": "./index.d.ts",
   "main": "",
   "scripts": {
     "check": "gts check",
@@ -12,7 +12,7 @@
     "prepare": "npm run compile",
     "pretest": "npm run compile",
     "posttest": "npm run check",
-    "test":"echo 'compilation worked!'"
+    "test": "echo 'compilation worked!'"
   },
   "keywords": [],
   "author": "Chris Dickinson <chris@neversaw.us> (http://neversaw.us/)",
