+++ tsconfig.json
@@ -5,6 +5,7 @@
     "outDir": "build"
   },
   "include": [
+    "index.d.ts",
     "src/*.ts",
     "src/**/*.ts",
     "test/*.ts",
